Fix casting

+++ esphome/components/m5-hbridge/hbridge.cpp
@@ -60,7 +60,14 @@ bool HBridgeComponent::write_driver_config_() {
 
   ESP_LOGD(TAG, "Setting motor to %d %s", duty_, ENUM_TO_DIRECTION_STRING[dir_].c_str());
 
-  uint8_t value[] = {dir_, 0, duty_ % 256, duty_ / 256, frequency_ % 256, frequency_ / 256};
+  uint8_t value[] = {
+    dir_,
+    0,
+    (uint8_t)(duty_ % 256),
+    (uint8_t)(duty_ / 256),
+    (uint8_t)(frequency_ % 256),
+    (uint8_t)(frequency_ / 256)
+  };
   if ((this->last_error_ = this->write_register(HBridge_DRIVER_CONFIG, value, 6, true)) != esphome::i2c::ERROR_OK) {
     this->status_set_warning();
     ESP_LOGE(TAG, "write_register_(): I2C I/O error: %d", (int) this->last_error_);
